Hand merged the following files:

U chrome/src/nsChromeRegistry.cpp
U chrome/src/nsChromeRegistry.h
U content/base/src/nsFrameLoader.cpp
U content/base/src/nsFrameLoader.h
U dom/src/geolocation/nsGeolocation.cpp
U dom/src/geolocation/nsGeolocation.h
U netwerk/build/Makefile.in
U netwerk/cookie/nsCookieService.cpp
U netwerk/cookie/nsCookieService.h

+++ js/src/jsapi-tests/tests.cpp
@@ -66,7 +66,7 @@ int main(int argc, char *argv[])
             JSAPITestString messages = test->messages();
             printf("%s | %s | %.*s\n",
                    (test->knownFail ? "TEST-KNOWN-FAIL" : "TEST-UNEXPECTED-FAIL"),
-                   name, messages.length(), messages.begin());
+                   name, (int) messages.length(), messages.begin());
             if (!test->knownFail)
                 failures++;
         }
